Tools/dinov2-embeddings/requirements.txt
torch>=2.1
torchvision>=0.16
pillow>=10
numpy>=1.26
